# docs/03_API仕様/openapi.yaml
openapi: 3.1.0
info:
  title: 人事考課システム API
  description: |
    トキワテック人事考課システムのREST API仕様書。

    ## 認証
    Cloudflare Access (JWT) による認証を使用します。
    全てのAPIリクエストには `Authorization: Bearer <token>` ヘッダーが必要です。

    ## 共通仕様
    - 日時はISO 8601形式 (例: `2026-01-24T10:00:00Z`)
    - IDはUUID v4形式
    - ページネーションはoffset/limit方式
  version: 1.0.0
  contact:
    name: トキワテック開発チーム
    email: dev@tokiwa-tech.com
  license:
    name: Proprietary
    identifier: LicenseRef-proprietary
servers:
  - url: https://evaluation.tokiwa-tech.com/api/v1
    description: 本番環境
  - url: https://evaluation-staging.tokiwa-tech.com/api/v1
    description: ステージング環境
  - url: http://localhost:8787/api/v1
    description: ローカル開発環境
...
paths:
...
  /auth/logout:
    post:
      tags: [Auth]
      summary: ログアウト
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ログアウト成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /auth/me:
    get:
      tags: [Auth]
      summary: 現在のユーザー情報取得
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ユーザー情報
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

  # ============================================
  # 評価期間 API (5 endpoints)
  # ============================================
...
  /admin/settings:
    get:
      tags: [Admin]
      summary: システム設定取得
      operationId: getSettings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: システム設定
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'

    put:
      tags: [Admin]
      summary: システム設定更新
      operationId: updateSettings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSettingsRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
  /admin/statistics:
    get:
      tags: [Admin]
      summary: システム統計取得
      operationId: getStatistics
      security:
        - BearerAuth: []
      responses:
        '200':
          description: システム統計
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatisticsResponse'
  /admin/export:
    post:
      tags: [Admin, Export]
      summary: エクスポートジョブ開始
      operationId: startExportJob
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '202':
          description: ジョブ開始（非同期）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportJobResponse'
  /admin/export/{jobId}:
    get:
      tags: [Admin, Export]
      summary: エクスポートジョブ状態取得
      operationId: getExportJobStatus
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ジョブ状態
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportJobStatusResponse'
  /admin/audit-logs:
    get:
      tags: [Admin]
      summary: 監査ログ取得
      operationId: listAuditLogs
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: query
          schema:
            type: string
        - name: entityType
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 監査ログ一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogListResponse'

  # ============================================
  # ワークフロー管理 API (2 endpoints)
  # ============================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cloudflare Access JWT token
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: ページ番号
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: 1ページあたりの件数
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        hasNext:
          type: boolean

    # ============================================
    # 列挙型
    # ============================================
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [EMPLOYEE, MANAGER, HR, ADMIN]
        departmentId:
          type: string
        gradeId:
          type: string
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/User'

    # ============================================
    # 評価期間
    # ============================================
...
    AuditLogListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              userId:
                type: string
              action:
                type: string
              entityType:
                type: string
              entityId:
                type: string
              oldValue:
                type: object
              newValue:
                type: object
              createdAt:
                type: string
                format: date-time
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ============================================
    # ワークフロー
    # ============================================
...
    SettingsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            evaluationScoreRange:
              type: object
              properties:
                min:
                  type: integer
                max:
                  type: integer
            peerEvaluationLimit:
              type: integer
            reminderDays:
              type: array
              items:
                type: integer
            notificationChannels:
              type: array
              items:
                type: string
            allowSelfPeerSelection:
              type: boolean
            requireComment:
              type: boolean
            autoSubmitOnDeadline:
              type: boolean
    UpdateSettingsRequest:
      type: object
      properties:
        evaluationScoreRange:
          type: object
          properties:
            min:
              type: integer
            max:
              type: integer
        peerEvaluationLimit:
          type: integer
        reminderDays:
          type: array
          items:
            type: integer
        notificationChannels:
          type: array
          items:
            type: string
        allowSelfPeerSelection:
          type: boolean
        requireComment:
          type: boolean
        autoSubmitOnDeadline:
          type: boolean
    StatisticsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            totalEmployees:
              type: integer
            totalPeriods:
              type: integer
            totalCycles:
              type: integer
            totalScenarios:
              type: integer
            totalCurriculums:
              type: integer
            activePeriod:
              type: object
              nullable: true
              properties:
                id:
                  type: string
                name:
                  type: string
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
    ExportRequest:
      type: object
      properties:
        format:
          type: string
          enum: [csv, excel, json]
        entities:
          type: array
          items:
            type: string
        filters:
          type: object
    ExportJobResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            jobId:
              type: string
            status:
              type: string
            format:
              type: string
            entities:
              type: array
              items:
                type: string
            estimatedCompletionTime:
              type: string
              format: date-time
    ExportJobStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            jobId:
              type: string
            status:
              type: string
              enum: [PROCESSING, COMPLETED, FAILED]
            progress:
              type: integer
            downloadUrl:
              type: string
              nullable: true
            expiresAt:
              type: string
              format: date-time
              nullable: true

    # ============================================
    # Scenarios - 設問
    # ============================================
...